Made ProgressWindow.py into a base class so can be derived and used in other places instead of only the OpenPatientWindow.py
